requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.5
django-allauth==0.45.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
zeep==4.0.0
